commonizing logging into one spot
previously the logging was per file, this moves to a common logger that
each file now includes as well as introduces debug logging to actually
make it possible to track down extremely annoying items.

logs are also now persisted across reloads so you can see what happened
and if something actually changed. control is via the env files
otherwise it defaults to the previous behavior.